requirements.txt
boto3
streamlit
pillow
langchain
python-dotenv
